refactor: extract in a strategy interface token types retrieval
